6.0.0.4: attempt to copy gnome-mimeapps.list to missing defaults.list
In the past, defaults.list was provided by desktop-file-utils,
but in 24.04 it isn't.

cinnamon-desktop-data adds a minimal x-cinnamon-mimeapps.list.
LMDE 6 only has a defaults.list (not coming from any deb).
Mint 22 only has a mimeapps.list (also not coming from any deb).

So I don't see any "proper" way to solve this.
Since we base on Ubuntu, copy the gnome-mimeapps.list
which should always be available in our context.

I noticed that after a reboot, the system automatically
changed the list, substituting installed apps for ones
that I presume are not installed on wasta-linux.